Refactors login to Ducks
