--- client/actions/dice_actions.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const DICE_ADDED = 'DICE_ADDED';
-const DICE_CLEARED = 'DICE_CLEARED';
-const DICE_ROLLED = 'DICE_ROLLED';
-
-module.exports = {
-  DICE_ADDED,
-  DICE_CLEARED,
-  DICE_ROLLED,
-};
