--- client/actions/login_actions.js
@@ -1,56 +1,0 @@
-'use strict';
-
-/**
- * Login
- */
-const LOGIN_START = 'LOGIN_START';
-
-/**
- *
- */
-const LOGIN_SUCCESSFUL = 'LOGIN_SUCCESSFUL';
-
-/**
- *
- */
-const LOGIN_FAILURE = 'LOGIN_FAILURE';
-
-const REGISTER_START = 'REGISTER_START';
-const REGISTER_SUCCESS = 'REGISTER_SUCCESS';
-const REGISTER_FAIL = 'REGISTER_FAIL';
-
-const LOGOUT_START = 'LOGOUT_START';
-const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-const LOGOUT_FAILURE = 'LOGOUT_FAILURE';
-
-const SYNCHRONIZE_LOGIN_STATE_START = 'SYNCHRONIZE_LOGIN_STATE_START';
-const SYNCHRONIZE_LOGIN_STATE_SUCCESS = 'SYNCHRONIZE_LOGIN_STATE_SUCCESS';
-const SYNCHRONIZE_LOGIN_STATE_FAILURE = 'SYNCHRONIZE_LOGIN_STATE_FAILURE';
-
-const REQUEST_SLACK_TOKEN_START = 'REQUEST_SLACK_TOKEN_START';
-const REQUEST_SLACK_TOKEN_SUCCESS = 'REQUEST_SLACK_TOKEN_SUCCESS';
-const REQUEST_SLACK_TOKEN_FAIL = 'REQUEST_SLACK_TOKEN_FAIL';
-const DISMISS_SLACK_TOKEN = 'DISMISS_SLACK_TOKEN';
-
-module.exports = {
-  LOGIN_START,
-  LOGIN_SUCCESSFUL,
-  LOGIN_FAILURE,
-
-  REGISTER_START,
-  REGISTER_SUCCESS,
-  REGISTER_FAIL,
-
-  LOGOUT_START,
-  LOGOUT_SUCCESS,
-  LOGOUT_FAILURE,
-
-  SYNCHRONIZE_LOGIN_STATE_START,
-  SYNCHRONIZE_LOGIN_STATE_SUCCESS,
-  SYNCHRONIZE_LOGIN_STATE_FAILURE,
-
-  REQUEST_SLACK_TOKEN_START,
-  REQUEST_SLACK_TOKEN_SUCCESS,
-  REQUEST_SLACK_TOKEN_FAIL,
-  DISMISS_SLACK_TOKEN,
-};
